Jeremiah

+++ src/main/java/com/codeup/pressd/controllers/CommentController.java
@@ -63,8 +63,7 @@ public class CommentController {
         userComments.add(comment);
         workout.setComments(workoutComments);
         dbUser.setComments(userComments);
-        userDao.save(dbUser);
-        workoutDao.save(workout);
+//        workoutDao.save(workout);
         commentDao.save(comment);
         return "redirect:/workouts/"+id+"/comments";
     }
@@ -92,11 +91,15 @@ public class CommentController {
         return "redirect:/workouts";
     }
 
+
     @PostMapping("/comments/{id}/delete")
     public String deleteComment(@PathVariable Long id) {
-        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        User currentUser = userDao.getOne(user.getId());
-        commentDao.deleteById(id);
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Comment comment = commentDao.getOne(id);
+        if (currentUser.getId() == comment.getUser().getId()) {
+            commentDao.deleteById(id);
+        }
         return "redirect:/workouts";
+
     }
 }
